external_tools/requirements.txt
# Requirements for external_tools scripts
mediapipe
opencv-python
openai-whisper
torch
numpy
gentle 